refactor: apply review
