+++ src/brushes/basic.ts
@@ -1,8 +1,16 @@
-import { ClipRectAreaModel, PathRectModel, CircleModel, CircleStyle } from '@t/components/series';
+import {
+  ClipRectAreaModel,
+  PathRectModel,
+  CircleModel,
+  CircleStyle,
+  RectModel,
+  RectStyle,
+} from '@t/components/series';
 import { makeStyleObj } from '@src/helpers/style';
 import { LineModel } from '@t/components/axis';
 
 export type CircleStyleName = 'default' | 'hover';
+export type RectStyleName = 'default';
 
 const circleStyle = {
   default: {
@@ -90,4 +98,28 @@ export function line(ctx: CanvasRenderingContext2D, lineModel: LineModel) {
   ctx.lineTo(x2, y2);
   ctx.stroke();
   ctx.closePath();
+}
+
+export function rect(ctx: CanvasRenderingContext2D, model: RectModel) {
+  const { x, y, width, height, style, thickness = 0, color, borderColor = '#fff' } = model;
+
+  ctx.beginPath();
+
+  if (style) {
+    const styleObj = makeStyleObj<RectStyle, RectStyleName>(style, {});
+
+    Object.keys(styleObj).forEach((key) => {
+      ctx[key] = styleObj[key];
+    });
+  }
+
+  if (thickness) {
+    ctx.fillStyle = borderColor;
+    ctx.fillRect(x - thickness, y - thickness, width + thickness * 2, height + thickness * 2);
+    ctx.shadowColor = 'transparent';
+  }
+
+  ctx.fillStyle = color;
+  ctx.rect(x, y, width, height);
+  ctx.fill();
 }
