+++ src/charts/bubbleChart.ts
@@ -18,7 +18,6 @@ import * as basicBrushes from '@src/brushes/basic';
 import * as axisBrushes from '@src/brushes/axis';
 import * as tooltipBrush from '@src/brushes/tooltip';
 import * as circleLegendBrush from '@src/brushes/circleLegend';
-import * as boxSeriesBrush from '@src/brushes/boxSeries';
 import * as legendBrush from '@src/brushes/legend';
 import * as labelBrush from '@src/brushes/label';
 import * as exportMenuBrush from '@src/brushes/exportMenu';
@@ -65,7 +64,6 @@ export default class BubbleChart extends Chart<BaseOptions> {
       axisBrushes,
       tooltipBrush,
       circleLegendBrush,
-      boxSeriesBrush,
       legendBrush,
       labelBrush,
       exportMenuBrush,
