+++ src/charts/columnChart.ts
@@ -22,7 +22,6 @@ import HoveredSeries from '@src/component/hoveredSeries';
 
 import * as basicBrushes from '@src/brushes/basic';
 import * as axisBrushes from '@src/brushes/axis';
-import * as boxBrushes from '@src/brushes/boxSeries';
 import * as tooltipBrushes from '@src/brushes/tooltip';
 import * as legendBrush from '@src/brushes/legend';
 import * as labelBrush from '@src/brushes/label';
@@ -72,7 +71,6 @@ export default class ColumnChart extends Chart<ColumnChartOptions> {
     this.painter.addGroups([
       basicBrushes,
       axisBrushes,
-      boxBrushes,
       tooltipBrushes,
       legendBrush,
       labelBrush,
