+++ stories/data.ts
@@ -309,10 +309,10 @@ export const coordinateData = {
       name: 'SiteB',
       data: [
         { x: 1, y: 312 },
-        // { x: 3, y: 320 },
+        { x: 3, y: 320 },
         { x: 5, y: 295 },
-        // { x: 7, y: 300 },
-        // { x: 9, y: 320 },
+        { x: 7, y: 300 },
+        { x: 9, y: 320 },
         { x: 11, y: 30 },
         { x: 13, y: 20 },
       ],
